python-requirements.txt
tweepy
RPi.GPIO
gpiozero
picamera2
